sending force through the sphere
smashing the ground on the sphere will cause it to jerk in that
direction, until it eventually resettles back in the center. This is
just a prototype of the concept. For actual implementation, Id have to
revamp the spherical movement system in general to match the current
rigidbody system we have